io.c: fix clang -Werror,-Wshorten-64-to-32 on Linux sendfile
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@66099 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ io.c
@@ -11114,7 +11114,7 @@ nogvl_copy_stream_sendfile(struct copy_stream_struct *stp)
         }
         stp->syserr = "sendfile";
         stp->error_no = errno;
-        return ss;
+        return (int)ss;
     }
     return 1;
 }
